requirements.txt
streamlit==1.33.0
streamlit-option-menu==0.3.6
google-generativeai==0.4.0
Pillow==10.2.0
